requirements.txt
# Core dependencies for production
streamlit==1.45.1
openai==1.82.0
python-dotenv==1.0.1
llama-parse==0.6.23
requests==2.32.3
streamlit-option-menu==0.4.0
Pillow>=10.0.0  # For image processing (PIL)
numpy>=1.24.0,<2.1.0  
tqdm==4.67.1
# AWS SDK (for S3 upload and future AWS services)
boto3>=1.34.0
# Web search functionality
ddgs>=1.0.0